The proper map is rendered instead
